Moved candlestick series to subfolder.
