+++ js/Series/Candlestick/CandlestickSeries.js
@@ -21,15 +21,15 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../Core/Series/Series.js';
+import BaseSeries from '../../Core/Series/Series.js';
 var _a = BaseSeries.seriesTypes, ColumnSeries = _a.column, OHLCSeries = _a.ohlc;
 var columnProto = ColumnSeries.prototype;
-import O from '../Core/Options.js';
+import O from '../../Core/Options.js';
 var defaultOptions = O.defaultOptions;
-import palette from '../Core/Color/Palette.js';
-import U from '../Core/Utilities.js';
+import palette from '../../Core/Color/Palette.js';
+import U from '../../Core/Utilities.js';
 var extend = U.extend, merge = U.merge;
-import './OHLC/OHLCSeries.js';
+import '../OHLC/OHLCSeries.js';
 /* *
  *
  *  Code
